DAC6-3892 | refactor it packages
