+++ it/test/controllers/IndexControllerISpec.scala
@@ -21,9 +21,9 @@ import play.api.http.Status._
 import play.api.libs.ws.{DefaultWSCookie, WSClient}
 import play.api.mvc.{Session, SessionCookieBaker}
 import play.api.test.Helpers.{await, defaultAwaitTimeout}
-import utils.{AuthStubs, ISpecBase}
+import utils.ISpecBase
 
-class IndexControllerISpec extends PlaySpec with ISpecBase with AuthStubs {
+class IndexControllerISpec extends PlaySpec with ISpecBase {
 
   lazy val wsClient: WSClient = app.injector.instanceOf[WSClient]
   val session                 = Session(Map("authToken" -> "abc123", "role" -> "admin"))
@@ -60,7 +60,7 @@ class IndexControllerISpec extends PlaySpec with ISpecBase with AuthStubs {
           |}
           |}""".stripMargin
 
-      stubPostResponseB(readSubscriptionUrl, OK, responseDetailString)
+      stubPostResponse(readSubscriptionUrl, OK, responseDetailString)
       val response = await(
         buildClient()
           .withFollowRedirects(false)
